fix: fetch батч мемстатса положить поближе к функции
